+++ exporters/prometheus/exporter_test.go
@@ -55,7 +55,36 @@ func TestPrometheusExporter(t *testing.T) {
 				counter, err := meter.Float64Counter(
 					"foo",
 					otelmetric.WithDescription("a simple counter"),
-					otelmetric.WithUnit("ms"),
+					otelmetric.WithUnit("s"),
+				)
+				require.NoError(t, err)
+				counter.Add(ctx, 5, opt)
+				counter.Add(ctx, 10.3, opt)
+				counter.Add(ctx, 9, opt)
+
+				attrs2 := attribute.NewSet(
+					attribute.Key("A").String("D"),
+					attribute.Key("C").String("B"),
+					attribute.Key("E").Bool(true),
+					attribute.Key("F").Int(42),
+				)
+				counter.Add(ctx, 5, otelmetric.WithAttributeSet(attrs2))
+			},
+		},
+		{
+			name:         "counter that already has the unit suffix",
+			expectedFile: "testdata/counter.txt",
+			recordMetrics: func(ctx context.Context, meter otelmetric.Meter) {
+				opt := otelmetric.WithAttributes(
+					attribute.Key("A").String("B"),
+					attribute.Key("C").String("D"),
+					attribute.Key("E").Bool(true),
+					attribute.Key("F").Int(42),
+				)
+				counter, err := meter.Float64Counter(
+					"foo.seconds",
+					otelmetric.WithDescription("a simple counter"),
+					otelmetric.WithUnit("s"),
 				)
 				require.NoError(t, err)
 				counter.Add(ctx, 5, opt)
@@ -85,7 +114,7 @@ func TestPrometheusExporter(t *testing.T) {
 				counter, err := meter.Float64Counter(
 					"foo",
 					otelmetric.WithDescription("a simple counter without a total suffix"),
-					otelmetric.WithUnit("ms"),
+					otelmetric.WithUnit("s"),
 				)
 				require.NoError(t, err)
 				counter.Add(ctx, 5, opt)
@@ -415,7 +444,7 @@ func TestMultiScopes(t *testing.T) {
 	fooCounter, err := provider.Meter("meterfoo", otelmetric.WithInstrumentationVersion("v0.1.0")).
 		Int64Counter(
 			"foo",
-			otelmetric.WithUnit("ms"),
+			otelmetric.WithUnit("s"),
 			otelmetric.WithDescription("meter foo counter"))
 	assert.NoError(t, err)
 	fooCounter.Add(ctx, 100, otelmetric.WithAttributes(attribute.String("type", "foo")))
@@ -423,7 +452,7 @@ func TestMultiScopes(t *testing.T) {
 	barCounter, err := provider.Meter("meterbar", otelmetric.WithInstrumentationVersion("v0.1.0")).
 		Int64Counter(
 			"bar",
-			otelmetric.WithUnit("ms"),
+			otelmetric.WithUnit("s"),
 			otelmetric.WithDescription("meter bar counter"))
 	assert.NoError(t, err)
 	barCounter.Add(ctx, 200, otelmetric.WithAttributes(attribute.String("type", "bar")))
@@ -571,7 +600,7 @@ func TestDuplicateMetrics(t *testing.T) {
 				bazA.Add(ctx, 100, withTypeBar)
 
 				bazB, err := meterB.Int64Counter("bar",
-					otelmetric.WithUnit("ms"),
+					otelmetric.WithUnit("s"),
 					otelmetric.WithDescription("meter bar"))
 				assert.NoError(t, err)
 				bazB.Add(ctx, 100, withTypeBar)
@@ -589,7 +618,7 @@ func TestDuplicateMetrics(t *testing.T) {
 				barA.Add(ctx, 100, withTypeBar)
 
 				barB, err := meterB.Int64UpDownCounter("bar",
-					otelmetric.WithUnit("ms"),
+					otelmetric.WithUnit("s"),
 					otelmetric.WithDescription("meter gauge bar"))
 				assert.NoError(t, err)
 				barB.Add(ctx, 100, withTypeBar)
@@ -607,7 +636,7 @@ func TestDuplicateMetrics(t *testing.T) {
 				barA.Record(ctx, 100, withAB)
 
 				barB, err := meterB.Int64Histogram("bar",
-					otelmetric.WithUnit("ms"),
+					otelmetric.WithUnit("s"),
 					otelmetric.WithDescription("meter histogram bar"))
 				assert.NoError(t, err)
 				barB.Record(ctx, 100, withAB)